Fix mypy errors
