+++ homeassistant/components/shelly/sensor.py
@@ -161,7 +161,7 @@ SENSORS: Final[dict[tuple[str, str], BlockAttributeDescription]] = {
         value=lambda value: round(value, 1),
         device_class=sensor.DEVICE_CLASS_TEMPERATURE,
         state_class=sensor.STATE_CLASS_MEASUREMENT,
-        available=lambda block: block.extTemp != 999,
+        available=lambda block: cast(bool, block.extTemp != 999),
     ),
     ("sensor", "humidity"): BlockAttributeDescription(
         name="Humidity",
@@ -169,7 +169,7 @@ SENSORS: Final[dict[tuple[str, str], BlockAttributeDescription]] = {
         value=lambda value: round(value, 1),
         device_class=sensor.DEVICE_CLASS_HUMIDITY,
         state_class=sensor.STATE_CLASS_MEASUREMENT,
-        available=lambda block: block.extTemp != 999,
+        available=lambda block: cast(bool, block.extTemp != 999),
     ),
     ("sensor", "luminosity"): BlockAttributeDescription(
         name="Luminosity",
